Fixed some bugs
After defining north adjustment angle was not set on LocationService. 
Async persistance task in ProjectSiteActivity does not require a ui any more, we also could save results in background
if a new WiFi scan is started, the old one is stopped first, otherwise we would get duplicate results.


git-svn-id: https://wificompass.googlecode.com/svn/trunk@176 6f3e661d-c991-e48d-a97c-fec7ba324fcc

+++ src/at/fhstp/wificompass/model/Project.java
@@ -194,9 +194,10 @@ public class Project extends BaseDaoEnabled<Project, Integer>implements XMLSeria
 	 */
 	@Override
 	public int delete() throws SQLException {
-		for(ProjectSite site: sites){
-			site.delete();
-		}
+		if(sites!=null)
+			for(ProjectSite site: sites){
+				site.delete();
+			}
 		return super.delete();
 	}
